Restructure folders and add tests
