+++ src/pages/editlist/hooks/tests/useEditListFormInformationState.test.ts
@@ -1,6 +1,6 @@
 import { renderHook, act } from '@testing-library/react-hooks';
 import { useEditListFormState } from '../useEditListState';
-import { STATUS_VALUES, VISIBILITY_VALUES } from '../../types';
+import { STATUS_VALUES, VISIBILITY_VALUES } from '../../../../interfaces';
 
 describe('useEditListFormState', () => {
   const initialValues = {
